requirements.txt
flask
Flask-Dance
Flask-BasicAuth
Flask-SQLAlchemy
gunicorn
numpy
pandas
psycopg2
python-dotenv
requests
